+++ crates/wzp-android/src/pipeline.rs
@@ -138,7 +138,7 @@ impl Pipeline {
             let is_repair = header.is_repair();
             if let Err(e) = self.fec_decoder.add_symbol(
                 header.fec_block as u8,
-                (header.fec_block >> 8) as u8,
+                header.fec_block >> 8,
                 is_repair,
                 &packet.payload,
             ) {
